for some reason everything is showing up as modified. Anyway, I updated Regular_triangulation to set the time and added double.h to the includes for the polynomial interval arithmetic support

+++ Packages/Kinetic_data_structures/include/CGAL/KDS/Two_list_pointer_event_queue.h
@@ -608,7 +608,7 @@ protected:
 
 
   unsigned int max_front_size() const {
-    return std::max(10U, static_cast<unsigned int>(std::sqrt(front_.size()+back_.size())));
+    return std::max(10U, static_cast<unsigned int>(std::sqrt(static_cast<double>(front_.size()+back_.size()))));
   }
 
   Queue front_, back_;
